tests: moved test to its own
